pipeline/requirements.txt
requests
boto3
pyyaml
python-dotenv
